+++ CHANGELOG.md
@@ -52,6 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * Added MaxItemsPerUser filter to remove users with extreme amounts of interactions from a dataframe.
     * Added the `SessionDataFramePreprocessor` which cuts user histories into sessions while processing data into InteractionMatrices.
 
+* __pipelines__
+    * Added `get_metrics_dataframe` function, which returns the metrics as a dataframe, rather than a dict.
+
 ## [0.2.2] - ![](https://img.shields.io/date/1643025589.svg?label=2022-01-24)
 
 ### Additions